test(sam2-hoi): own engine timing coverage
Move the TRTMC_ENGINE_TIMING construction and event-retention assertions from the shared TensorRT module test into the SAM2-HOI model-owned GPU test manifest. Preserve every existing timing case while allowing model impact analysis to route the coverage to the owning family.
